Refine image scaling in FullScreenImagePreview and fix coordinate types in ManualCameraScanView
- Add `currentImagePathState` to `FullScreenImagePreview` to track the active image and reset scale only when the path actually changes.
- Reset `scaleState` to `1f` when bitmap or viewport dimensions are updated to ensure a consistent initial view.
- Reference `scaleState.floatValue` directly within `detectTransformGestures` and `graphicsLayer` for better state consistency.
- Explicitly cast normalized scan area coordinates to `Float` in `ManualCameraScanView` to ensure type safety during UI calculations.

+++ app/src/main/java/com/sergio/barcodescanner/ManualCameraScanView.kt
@@ -430,10 +430,10 @@ fun ManualCameraScanView(
                     val offsetX = (viewW - srcW * scale) / 2f
                     val offsetY = (viewH - srcH * scale) / 2f
 
-                    val normalizedLeft = minOf(rect.left, rect.right)
-                    val normalizedTop = minOf(rect.top, rect.bottom)
-                    val normalizedRight = maxOf(rect.left, rect.right)
-                    val normalizedBottom = maxOf(rect.top, rect.bottom)
+                    val normalizedLeft = minOf(rect.left, rect.right).toFloat()
+                    val normalizedTop = minOf(rect.top, rect.bottom).toFloat()
+                    val normalizedRight = maxOf(rect.left, rect.right).toFloat()
+                    val normalizedBottom = maxOf(rect.top, rect.bottom).toFloat()
 
                     val left = normalizedLeft * scale + offsetX
                     val top = normalizedTop * scale + offsetY
